chore: separate zscore plot from helper fn
